Create KMeans Clustering for 2019 US Survey of Consumer Finances Data
I use data from the 2019 US Survey of Consumer Finances (https://www.federalreserve.gov/econres/scfindex.htm) to perform KMeans clustering (unsupervised learning) with implications for social stratification studies and customer segmentation studies. Further details of this dataset will be broken down shortly but for now, this dataset comprises of 351 columns of data collected from 28, 885 US households.

Firstly, I run EDA to explore the dataset and identify features which may be of interest. Then I use tests of variance (including trimmed variance) to pinpoint and select variables for our clustering model. This results in a 5-dimensional feature set which I use to train the KMeans model.

Differences in scale across the 5 features necessitate that I employ standardisation measures. Specifically, I use the ScikitLearn 'StandardScaler()' transformer as part of my model building pipeline.

Using a for-loop I build a series of KMeans models with differing number-of-clusters parameter settings and for each, Inertia and Silhouette scores are calculated and instrumental in selecting the best final model with optimum hyperparameters.

As part of this project I conduct Principal Component Analysis to reduce high dimensionality data with emphasis on visualisation.